defined transaction amount type as float

+++ database/migrations/2024_06_11_184548_create_transactions_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
             $table->string('description');
-            $table->integer('amount');
+            $table->float('amount');
             $table->timestamps();
         });
     }
